Moved commands to one class
- merged commands to one class
- permission changed to velocitab.command.reload
- changed the link for the wiki
